Re-organised property tests into individual pckages
