+++ grakn-test/test-integration/src/test/java/ai/grakn/test/property/kb/InsertQueryPropertyTest.java
@@ -16,18 +16,17 @@
  * along with Grakn. If not, see <http://www.gnu.org/licenses/gpl.txt>.
  */
 
-package ai.grakn.test.property;
+package ai.grakn.test.property.kb;
 
 import ai.grakn.GraknTx;
+import ai.grakn.exception.GraknTxOperationException;
 import ai.grakn.exception.GraqlQueryException;
 import ai.grakn.generator.GraknTxs.Open;
 import ai.grakn.graql.Graql;
 import ai.grakn.graql.admin.VarPatternAdmin;
 import ai.grakn.graql.admin.VarProperty;
 import ai.grakn.graql.internal.pattern.Patterns;
-import ai.grakn.graql.internal.pattern.property.PlaysProperty;
-import ai.grakn.graql.internal.pattern.property.RelatesProperty;
-import ai.grakn.graql.internal.pattern.property.SubProperty;
+import ai.grakn.graql.internal.pattern.property.IsaProperty;
 import com.pholser.junit.quickcheck.Property;
 import com.pholser.junit.quickcheck.generator.Size;
 import com.pholser.junit.quickcheck.runner.JUnitQuickcheck;
@@ -36,28 +35,40 @@ import org.junit.Rule;
 import org.junit.rules.ExpectedException;
 import org.junit.runner.RunWith;
 
+import java.util.Collections;
 import java.util.Set;
 
+import static org.hamcrest.MatcherAssert.assertThat;
+import static org.hamcrest.Matchers.not;
+import static org.hamcrest.collection.IsEmptyCollection.empty;
 import static org.junit.Assume.assumeFalse;
 
 @RunWith(JUnitQuickcheck.class)
-public class DefineQueryPropertyTest {
+public class InsertQueryPropertyTest {
 
     @Rule
     public ExpectedException exception = ExpectedException.none();
 
+    @Property
+    public void ifAPropertyCanBeInserted_AResultShouldBeReturned(@Open GraknTx tx, VarProperty property){
+        VarPatternAdmin pattern = Patterns.varPattern(Graql.var("x"), Collections.singleton(property));
+        try {
+            assertThat(tx.graql().insert(pattern).execute(), not(empty()));
+        } catch(GraqlQueryException | GraknTxOperationException e){
+            // IGNORED
+        }
+    }
+
     @Ignore("Currently no error message is returned when trying to insert an empty set of propoerties. I am not entirely sure this is correct")
     @Property
-    public void aDefineQueryWithoutASubOrPlaysOrRelatesProperty_CannotBeInserted(@Open GraknTx tx, @Size(max=5) Set<VarProperty> properties){
-        boolean containsSub = properties.stream().anyMatch(SubProperty.class::isInstance);
-        boolean containsPlays = properties.stream().anyMatch(PlaysProperty.class::isInstance);
-        boolean containsRelates = properties.stream().anyMatch(RelatesProperty.class::isInstance);
-        assumeFalse(containsSub || containsPlays || containsRelates);
+    public void anInsertQueryWithoutAnIsaProperty_CannotBeInserted(@Open GraknTx tx, @Size(max=5) Set<VarProperty> properties){
+        boolean containsIsa = properties.stream().anyMatch(IsaProperty.class::isInstance);
+        assumeFalse(containsIsa);
 
         VarPatternAdmin pattern = Patterns.varPattern(Graql.var("x"), properties);
 
         exception.expect(GraqlQueryException.class);
 
-        tx.graql().define(pattern).execute();
+        tx.graql().insert(pattern).execute();
     }
 }
